pyproject.toml
[project]
name = "semantiq-nb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alpaca-py>=0.43.2",
    "cufflinks>=0.17.3",
    "datasets>=4.3.0",
    "gensim>=4.4.0",
    "ipywidgets>=8.1.8",
    "load-dotenv>=0.1.0",
    "matplotlib>=3.10.7",
    "meteostat>=1.7.6",
    "nbformat>=5.10.4",
    "nltk>=3.9.2",
    "numpy>=2.3.4",
    "plotly>=6.4.0",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
